Extract source folder filter stuff to separate class
