+++ org.projectusus.ui.dependencygraph/src/org/projectusus/ui/dependencygraph/common/DependencyGraphView.java
@@ -55,7 +55,7 @@ import org.projectusus.ui.dependencygraph.filters.PackagenameNodeFilter;
 import org.projectusus.ui.dependencygraph.handlers.ChangeZoom;
 import org.projectusus.ui.dependencygraph.nodes.GraphNode;
 
-public abstract class DependencyGraphView extends ViewPart implements IRestrictNodesFilterProvider, IShowInTarget, IZoomableWorkbenchPart, IMenuListener {
+public abstract class DependencyGraphView extends ViewPart implements IRestrictNodesFilterProvider, IShowInTarget, IZoomableWorkbenchPart, IMenuListener, IRefreshable {
 
     private final DependencyGraphModel model;
     private DependencyGraphViewer graphViewer;
