+++ org.projectusus.ui.dependencygraph/src/org/projectusus/ui/dependencygraph/common/IRefreshable.java
@@ -1,0 +1,7 @@
+package org.projectusus.ui.dependencygraph.common;
+
+public interface IRefreshable {
+
+    void refresh();
+
+}
